+++ src/structlog/_config.py
@@ -23,7 +23,7 @@ from typing import (
 
 from ._log_levels import make_filtering_bound_logger
 from ._loggers import PrintLoggerFactory
-from .dev import ConsoleRenderer, _has_colorama, set_exc_info
+from .dev import ConsoleRenderer, _use_colors, set_exc_info
 from .processors import StackInfoRenderer, TimeStamper, add_log_level
 from .types import BindableLogger, Context, Processor, WrappedLogger
 
@@ -39,7 +39,7 @@ _BUILTIN_DEFAULT_PROCESSORS: Sequence[Processor] = [
     set_exc_info,
     TimeStamper(fmt="%Y-%m-%d %H:%M.%S", utc=False),
     ConsoleRenderer(
-        colors=_has_colorama and sys.stdout is not None and sys.stdout.isatty()
+        colors=_use_colors and sys.stdout is not None and sys.stdout.isatty()
     ),
 ]
 _BUILTIN_DEFAULT_CONTEXT_CLASS = cast(Type[Context], dict)